requirements.txt
ipykernel
selenium
webdriver_manager
Pillow
pywhatkit
ascii_magic